python/capture/requirements_capture.txt
chipwhisperer
argparse
numpy
time
tqdm
copy
pathlib
pyyaml